requirements.txt
streamlit
streamlit-aggrid
pandas
numpy
tensorflow
scikit-learn
fastdtw
tqdm
faiss-cpu